+++ R/IDConverter-package.R
@@ -5,7 +5,7 @@
 
 # Suppress R CMD check NOTE for data.table NSE variables
 utils::globalVariables(c(
-  ".", "alias", "case_id", "ensgene", "file_id_tumor", "file_id_normal",
+  ".", "alias", "case_id", "confidence", "ensgene", "file_id_tumor", "file_id_normal",
   "i.id", "i.type", "id", "normal_sample", "query",
   "sample_id", "symbol", "synonym", "tissue_type", "tumor_sample",
   "type", "type2"